+++ src/Walkers/WalkerWeighted02.java
@@ -138,7 +138,7 @@ public class WalkerWeighted02 extends Walker implements UnitInterface
     }
     private double heuristicGrad(int x, int y)
     {
-        double curGrad = (double)Math.abs(dest.getY() - y) / (double)Math.abs(dest.getX() - x);
+        double curGrad = Math.abs(dest.getY() - y) / Math.abs(dest.getX() - x);
         //System.out.println("CurGrad:"+curGrad+"| staGrad:"+this.grad);
         if (Double.isInfinite(Math.abs(curGrad - this.grad)))
         { return 0;
